Add entity parent and enemy classes
